Create Senor Cactus Intersect objects
Here is the function that allows two images to intersect each other in the game (i.e. the image of senor cactus intersects the image of any other object in the game). Basically you can use this to trigger the win/lose part of the game, the gain of life by touching the raindrop, etc...